# docs/reference/kumod.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "kumod",
    "description": "",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "2023.11.18-8b8f5312"
  },
  "paths": {
...
    "/api/admin/suspend/v1": {
      "get": {
        "tags": [
          "suspend"
        ],
        "summary": "List the active scheduled-queue suspensions",
        "description": "List the active scheduled-queue suspensions",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Suspended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuspendV1ListEntry"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "suspend"
        ],
        "summary": "Define a suspension for a scheduled queue",
        "description": "Define a suspension for a scheduled queue",
        "operationId": "suspend",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Suspended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuspendV1Response"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "suspend"
        ],
        "summary": "Remove a scheduled-queue suspension",
        "description": "Remove a scheduled-queue suspension",
        "operationId": "delete",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendV1CancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Removed the suspension"
          },
          "404": {
            "description": "Suspension either expired or was never valid"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SuspendV1CancelRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The id of the suspension to cancel"
          }
        }
      },
      "SuspendV1ListEntry": {
        "type": "object",
        "required": [
          "id",
          "reason",
          "duration"
        ],
        "properties": {
          "campaign": {
            "type": "string",
            "description": "The campaign name to match. If omitted, any campaign will match.",
            "nullable": true
          },
          "domain": {
            "type": "string",
            "description": "The domain name to match. If omitted, any domain will match.",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "description": "Specifies how long this suspension remains active."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The id of the suspension. This can be used later to cancel\nthe suspension."
          },
          "reason": {
            "type": "string",
            "description": "The reason for the suspension",
            "example": "pause while working on resolving a deliverability issue"
          },
          "tenant": {
            "type": "string",
            "description": "The tenant name to match. If omitted, any tenant will match.",
            "nullable": true
          }
        }
      },
      "SuspendV1Request": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "campaign": {
            "type": "string",
            "description": "The campaign name to match. If omitted, any campaign will match.",
            "nullable": true
          },
          "domain": {
            "type": "string",
            "description": "The domain name to match. If omitted, any domain will match.",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "description": "Specifies how long this suspension remains active.",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "description": "The reason for the suspension",
            "example": "pause while working on resolving a block with the destination postmaster"
          },
          "tenant": {
            "type": "string",
            "description": "The tenant name to match. If omitted, any tenant will match.",
            "nullable": true
          }
        }
      },
      "SuspendV1Response": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The id of the suspension. This can be used later to cancel\nthe suspension."
          }
        }
      }
    },
...
  }
}